+++ R/remotes.R
@@ -4,30 +4,33 @@
 #' Given a remote specification, resolve it into an `renv` package record that
 #' can be used for download and installation (e.g. with [install]).
 #'
-#' @param spec A remote specification.
+#' @param spec A remote specification. This should be a string, conforming
+#'   to the Remotes specification as defined in
+#'   <https://remotes.r-lib.org/articles/dependencies.html>.
 #'
 remote <- function(spec) {
   renv_scope_error_handler()
   renv_remotes_resolve(spec)
 }
 
-# take a short-form remotes entry, and generate a package record
-renv_remotes_resolve <- function(entry, latest = FALSE) {
+# take a short-form remotes spec, parse that into a remote,
+# and generate a corresopnding package record
+renv_remotes_resolve <- function(spec, latest = FALSE) {
 
-  # check for already-resolved lists
-  if (is.null(entry) || is.list(entry))
-    return(entry)
+  # check for already-resolved specs
+  if (is.null(spec) || is.list(spec))
+    return(spec)
 
   # check for URLs
-  if (grepl("^(?:file|https?)://", entry))
-    return(renv_remotes_resolve_url(entry, quiet = TRUE))
+  if (grepl("^(?:file|https?)://", spec))
+    return(renv_remotes_resolve_url(spec, quiet = TRUE))
 
   # check for paths to existing local files
-  first <- substring(entry, 1L, 1L)
-  local <- first %in% c("~", "/", ".") || renv_path_absolute(entry)
+  first <- substring(spec, 1L, 1L)
+  local <- first %in% c("~", "/", ".") || renv_path_absolute(spec)
 
   if (local) {
-    record <- catch(renv_remotes_resolve_path(entry))
+    record <- catch(renv_remotes_resolve_path(spec))
     if (!inherits(record, "error"))
       return(record)
   }
@@ -37,7 +40,7 @@ renv_remotes_resolve <- function(entry, latest = FALSE) {
 
     # build error message
     fmt <- "failed to resolve remote '%s'"
-    prefix <- sprintf(fmt, entry)
+    prefix <- sprintf(fmt, spec)
     message <- paste(prefix, e$message, sep = " -- ")
 
     # if the error occurred while running tests, and the error appears
@@ -55,36 +58,36 @@ renv_remotes_resolve <- function(entry, latest = FALSE) {
 
   # attempt the parse
   withCallingHandlers(
-    renv_remotes_resolve_impl(entry, latest),
+    renv_remotes_resolve_impl(spec, latest),
     error = error
   )
 
 }
 
-renv_remotes_resolve_impl <- function(entry, latest = FALSE) {
+renv_remotes_resolve_impl <- function(spec, latest = FALSE) {
 
-  parsed <- renv_remotes_parse(entry)
+  remote <- renv_remotes_parse(spec)
 
   resolved <- switch(
-    parsed$type,
-    bitbucket  = renv_remotes_resolve_bitbucket(parsed),
-    gitlab     = renv_remotes_resolve_gitlab(parsed),
-    github     = renv_remotes_resolve_github(parsed),
-    repository = renv_remotes_resolve_repository(parsed, latest),
-    git        = renv_remotes_resolve_git(parsed),
-    stopf("unknown remote type '%s'", parsed$type %||% "<NA>")
+    remote$type,
+    bitbucket  = renv_remotes_resolve_bitbucket(remote),
+    gitlab     = renv_remotes_resolve_gitlab(remote),
+    github     = renv_remotes_resolve_github(remote),
+    repository = renv_remotes_resolve_repository(remote, latest),
+    git        = renv_remotes_resolve_git(remote),
+    stopf("unknown remote type '%s'", remote$type %||% "<NA>")
   )
 
   reject(resolved, is.null)
 
 }
 
-renv_remotes_parse_impl <- function(entry, pattern, fields, perl = FALSE) {
+renv_remotes_parse_impl <- function(spec, pattern, fields, perl = FALSE) {
 
-  matches <- regexec(pattern, entry, perl = perl)
-  strings <- regmatches(entry, matches)[[1]]
+  matches <- regexec(pattern, spec, perl = perl)
+  strings <- regmatches(spec, matches)[[1]]
   if (empty(strings))
-    stopf("'%s' is not a valid remote", entry)
+    stopf("'%s' is not a valid remote", spec)
 
   if (length(fields) != length(strings))
     stop("internal error: field length mismatch in renv_remotes_parse_impl")
@@ -95,7 +98,7 @@ renv_remotes_parse_impl <- function(entry, pattern, fields, perl = FALSE) {
 
 }
 
-renv_remotes_parse_repos <- function(entry) {
+renv_remotes_parse_repos <- function(spec) {
 
   pattern <- paste0(
     "^",                               # start
@@ -105,12 +108,12 @@ renv_remotes_parse_repos <- function(entry) {
     "$"
   )
 
-  fields <- c("entry", "repository", "package", "version")
-  renv_remotes_parse_impl(entry, pattern, fields)
+  fields <- c("spec", "repository", "package", "version")
+  renv_remotes_parse_impl(spec, pattern, fields)
 
 }
 
-renv_remotes_parse_remote <- function(entry) {
+renv_remotes_parse_remote <- function(spec) {
 
   pattern <- paste0(
     "^",
@@ -123,17 +126,18 @@ renv_remotes_parse_remote <- function(entry) {
     "$"
   )
 
-  fields <- c("entry", "type", "host", "user", "repo", "subdir", "pull", "ref")
-  parsed <- renv_remotes_parse_impl(entry, pattern, fields)
+  fields <- c("spec", "type", "host", "user", "repo", "subdir", "pull", "ref")
+  remote <- renv_remotes_parse_impl(spec, pattern, fields)
 
-  if (!nzchar(parsed$repo))
-    stopf("'%s' is not a valid remote", entry)
+  if (!nzchar(remote$repo))
+    stopf("'%s' is not a valid remote", spec)
 
-  renv_remotes_parse_finalize(parsed)
+  renv_remotes_parse_finalize(remote)
 
 }
 
-renv_remotes_parse_git <- function(entry) {
+renv_remotes_parse_git <- function(spec) {
+
   pattern <- paste0(
     "^",
     "(?:([^@:]+)::)?",  # optional prefix, providing type
@@ -157,7 +161,7 @@ renv_remotes_parse_git <- function(entry) {
   )
 
   fields <- c(
-    "entry",
+    "spec",
       "type",
       "url",
         "protocol", "login", "host",
@@ -165,82 +169,83 @@ renv_remotes_parse_git <- function(entry) {
       "subdir",
       "pull",
       "ref"
-    )
-  parsed <- renv_remotes_parse_impl(entry, pattern, fields, perl = TRUE)
+  )
 
-  if (!nzchar(parsed$repo))
-    stopf("'%s' is not a valid remote", entry)
+  remote <- renv_remotes_parse_impl(spec, pattern, fields, perl = TRUE)
+  if (!nzchar(remote$repo))
+    stopf("'%s' is not a valid remote", spec)
 
   # If type has not been found & repo looks like a git repo, set it as git
   # (note that this parser also accepts entries which are not truly git
   # references, so we try to "fix up" after the fact)
-  if ("git" %in% c(parsed$type, parsed$ext, parsed$protocol))
-    parsed$type <- tolower(parsed$type %||% "git")
+  if ("git" %in% c(remote$type, remote$ext, remote$protocol))
+    remote$type <- tolower(remote$type %||% "git")
 
-  renv_remotes_parse_finalize(parsed)
+  renv_remotes_parse_finalize(remote)
 }
 
-renv_remotes_parse_finalize <- function(parsed) {
+renv_remotes_parse_finalize <- function(remote) {
+
   # default remote type is github
-  parsed$type <- tolower(parsed$type %||% "github")
+  remote$type <- tolower(remote$type %||% "github")
 
   # custom finalization for different remote types
   case(
-    parsed$type == "github" ~ renv_remotes_parse_finalize_github(parsed),
-    TRUE                    ~ parsed
+    remote$type == "github" ~ renv_remotes_parse_finalize_github(remote),
+    TRUE                    ~ remote
   )
 
 }
 
-renv_remotes_parse_finalize_github <- function(parsed) {
+renv_remotes_parse_finalize_github <- function(remote) {
 
   # split repo spec into pieces
-  repo <- parsed$repo %||% ""
+  repo <- remote$repo %||% ""
   parts <- strsplit(repo, "/", fixed = TRUE)[[1]]
   if (length(parts) < 2)
-    return(parsed)
+    return(remote)
 
   # form subdir from tail of repo
-  parsed$repo   <- paste(head(parts, n = 1L),  collapse = "/")
-  parsed$subdir <- paste(tail(parts, n = -1L), collapse = "/")
+  remote$repo   <- paste(head(parts, n = 1L),  collapse = "/")
+  remote$subdir <- paste(tail(parts, n = -1L), collapse = "/")
 
   # return modified remote
-  parsed
+  remote
 
 }
 
-renv_remotes_parse <- function(entry) {
+renv_remotes_parse <- function(spec) {
 
-  parsed <- catch(renv_remotes_parse_repos(entry))
-  if (!inherits(parsed, "error")) {
-    parsed$type <- "repository"
-    return(parsed)
+  remote <- catch(renv_remotes_parse_repos(spec))
+  if (!inherits(remote, "error")) {
+    remote$type <- "repository"
+    return(remote)
   }
 
-  parsed <- catch(renv_remotes_parse_remote(entry))
-  if (!inherits(parsed, "error")) {
-    parsed$type <- parsed$type %||% "github"
-    return(parsed)
+  remote <- catch(renv_remotes_parse_remote(spec))
+  if (!inherits(remote, "error")) {
+    remote$type <- remote$type %||% "github"
+    return(remote)
   }
 
-  parsed <- catch(renv_remotes_parse_git(entry))
-  if (!inherits(parsed, "error")) {
-    parsed$type <- parsed$type %||% "git"
-    return(parsed)
+  remote <- catch(renv_remotes_parse_git(spec))
+  if (!inherits(remote, "error")) {
+    remote$type <- remote$type %||% "git"
+    return(remote)
   }
 
-  stopf("failed to parse remote spec '%s'", entry)
+  stopf("failed to parse remote spec '%s'", spec)
 
 }
 
-renv_remotes_resolve_bitbucket <- function(entry) {
+renv_remotes_resolve_bitbucket <- function(remote) {
 
-  user   <- entry$user
-  repo   <- entry$repo
-  subdir <- entry$subdir
-  ref    <- entry$ref %||% getOption("renv.bitbucket.default_branch", "master")
+  user   <- remote$user
+  repo   <- remote$repo
+  subdir <- remote$subdir
+  ref    <- remote$ref %||% getOption("renv.bitbucket.default_branch", "master")
 
-  host <- entry$host %||% config$bitbucket.host()
+  host <- remote$host %||% config$bitbucket.host()
 
   # get commit sha for ref
   fmt <- "%s/repositories/%s/%s/commit/%s"
@@ -276,18 +281,18 @@ renv_remotes_resolve_bitbucket <- function(entry) {
 
 }
 
-renv_remotes_resolve_repository <- function(entry, latest) {
+renv_remotes_resolve_repository <- function(remote, latest) {
 
-  package <- entry$package
+  package <- remote$package
   if (package %in% renv_packages_base())
     return(renv_remotes_resolve_base(package))
 
-  version <- entry$version
-  repository <- entry$repository
+  version <- remote$version
+  repository <- remote$repository
 
   if (latest && is.null(version)) {
-    entry <- renv_available_packages_latest(package)
-    version <- entry$Version
+    remote <- renv_available_packages_latest(package)
+    version <- remote$Version
   }
 
   list(
@@ -323,7 +328,7 @@ renv_remotes_resolve_github_sha_pull <- function(host, user, repo, pull) {
 
 renv_remotes_resolve_github_sha_ref <- function(host, user, repo, ref) {
 
-  # build url for github endpoint
+  # build url for github commits endpoint
   fmt <- "%s/repos/%s/%s/commits/%s"
   origin <- renv_retrieve_origin(host)
   ref <- ref %||% getOption("renv.github.default_branch", default = "master")
@@ -414,23 +419,28 @@ renv_remotes_resolve_github_ref_impl <- function(host, user, repo) {
 
 }
 
-renv_remotes_resolve_github <- function(entry) {
+renv_remotes_resolve_github <- function(remote) {
 
   # resolve the reference associated with this repository
-  host <- entry$host %||% config$github.host()
-  user <- entry$user
-  repo <- entry$repo
-  ref  <- entry$ref %||% renv_remotes_resolve_github_ref(host, user, repo)
+  host <- remote$host %||% config$github.host()
+  user <- remote$user
+  repo <- remote$repo
+  spec <- remote$spec
+  ref  <- remote$ref %||% renv_remotes_resolve_github_ref(host, user, repo)
+
+  # handle '*release' refs
+  if (identical(ref, "*release"))
+    ref <- renv_remotes_resolve_github_release(host, user, repo, spec)
 
   # resolve the sha associated with the ref / pull
-  pull   <- entry$pull %||% ""
+  pull   <- remote$pull %||% ""
   sha <- case(
     nzchar(pull) ~ renv_remotes_resolve_github_sha_pull(host, user, repo, pull),
     nzchar(ref)  ~ renv_remotes_resolve_github_sha_ref(host, user, repo, ref)
   )
 
   # read DESCRIPTION
-  subdir <- entry$subdir
+  subdir <- remote$subdir
   desc <- renv_remotes_resolve_github_description(host, user, repo, subdir, sha)
 
   list(
@@ -448,15 +458,46 @@ renv_remotes_resolve_github <- function(entry) {
 
 }
 
-renv_remotes_resolve_git <- function(parsed) {
+renv_remotes_resolve_github_release <- function(host, user, repo, spec) {
 
-  package <- parsed$repo
-  url <- parsed$url
-  subdir <- parsed$subdir
+  # build url for github releases endpoint
+  fmt <- "%s/repos/%s/%s/releases?per_page=1"
+  origin <- renv_retrieve_origin(host)
+  url <- sprintf(fmt, origin, user, repo)
+
+  # prepare headers
+  headers <- c(Accept = "application/vnd.github.v3+json")
+
+  # make request to endpoint
+  releases <- renv_scope_tempfile("renv-releases-")
+  download(
+    url      = url,
+    destfile = releases,
+    type     = "github",
+    quiet    = TRUE,
+    headers  = headers
+  )
+
+  # get reference associated with this tag
+  json <- renv_json_read(releases)
+  if (empty(json)) {
+    fmt <- "could not find any releases associated with remote '%s'"
+    stopf(fmt, sub("[*]release$", "", spec))
+  }
+
+  json[[1L]][["tag_name"]]
+
+}
+
+renv_remotes_resolve_git <- function(remote) {
+
+  package <- remote$repo
+  url     <- remote$url
+  subdir  <- remote$subdir
 
   # handle git ref
-  pull <- parsed$pull %||% ""
-  ref  <- parsed$ref %||% ""
+  pull <- remote$pull %||% ""
+  ref  <- remote$ref %||% ""
 
   # resolve ref from pull if set
   if (nzchar(pull))
@@ -571,14 +612,14 @@ renv_remotes_resolve_gitlab_ref_impl <- function(host, user, repo) {
 
 }
 
-renv_remotes_resolve_gitlab <- function(entry) {
+renv_remotes_resolve_gitlab <- function(remote) {
 
-  host   <- entry$host %||% config$gitlab.host()
-  user   <- entry$user
-  repo   <- entry$repo
-  subdir <- entry$subdir %||% ""
+  host   <- remote$host %||% config$gitlab.host()
+  user   <- remote$user
+  repo   <- remote$repo
+  subdir <- remote$subdir %||% ""
 
-  ref <- entry$ref %||% renv_remotes_resolve_gitlab_ref(host, user, repo)
+  ref <- remote$ref %||% renv_remotes_resolve_gitlab_ref(host, user, repo)
 
   parts <- c(if (nzchar(subdir)) subdir, "DESCRIPTION")
   descpath <- URLencode(paste(parts, collapse = "/"), reserved = TRUE)
@@ -620,18 +661,18 @@ renv_remotes_resolve_gitlab <- function(entry) {
 
 }
 
-renv_remotes_resolve_url <- function(entry, quiet = FALSE) {
+renv_remotes_resolve_url <- function(url, quiet = FALSE) {
 
   tempfile <- renv_scope_tempfile("renv-url-")
-  writeLines(entry, con = tempfile)
+  writeLines(url, con = tempfile)
   hash <- tools::md5sum(tempfile)
 
-  ext <- fileext(entry, default = ".tar.gz")
+  ext <- fileext(url, default = ".tar.gz")
   name <- paste(hash, ext, sep = "")
   path <- renv_paths_source("url", name)
 
   ensure_parent_directory(path)
-  download(entry, path, quiet = quiet)
+  download(url, path, quiet = quiet)
 
   desc <- renv_description_read(path)
 
@@ -640,26 +681,26 @@ renv_remotes_resolve_url <- function(entry, quiet = FALSE) {
     Version    = desc$Version,
     Source     = "URL",
     RemoteType = "url",
-    RemoteUrl  = entry,
+    RemoteUrl  = url,
     Path       = path
   )
 
 }
 
-renv_remotes_resolve_path <- function(entry) {
+renv_remotes_resolve_path <- function(path) {
 
   # check for existing path
-  path <- renv_path_normalize(entry, winslash = "/", mustWork = TRUE)
+  norm <- renv_path_normalize(path, winslash = "/", mustWork = TRUE)
 
   # first, check for a common extension
-  if (renv_archive_type(entry) %in% c("tar", "zip"))
-    return(renv_remotes_resolve_path_impl(path))
+  if (renv_archive_type(path) %in% c("tar", "zip"))
+    return(renv_remotes_resolve_path_impl(norm))
 
   # otherwise, if this is the path to a package project, use the sources as-is
-  if (renv_project_type(path) == "package")
-    return(renv_remotes_resolve_path_impl(path))
+  if (renv_project_type(norm) == "package")
+    return(renv_remotes_resolve_path_impl(norm))
 
-  stopf("there is no package at path '%s'", entry)
+  stopf("there is no package at path '%s'", path)
 
 }