+++ tests/testthat/test-retrieve.R
@@ -344,3 +344,10 @@ test_that("we respect the default branch for gitlab repositories", {
   remote <- renv_remotes_resolve("gitlab::kevinushey/main")
   expect_equal(remote$RemoteRef, "main")
 })
+
+test_that("renv can retrieve the latest release associated with a project", {
+  skip_on_cran()
+  skip_on_ci()
+  remote <- renv_remotes_resolve("rstudio/keras@*release")
+  expect_true(is.list(remote))
+})